+++ telemonitor/extensions/systemd_service/files/telemonitor_start.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+
 cd "$(dirname "$0")" # Change workdir to shell script location
 cd ../../../..       # Change workdir to the root of Telemonitor project